Adding and moving files around for the scaffolfing with Firebase context
